+++ src/Encoder/FloatEncoder.php
@@ -10,9 +10,12 @@ namespace Riimu\Kit\PHPEncoder\Encoder;
  */
 class FloatEncoder implements Encoder
 {
+    /** The maximum value that can be accurately represented by a float */
+    const FLOAT_MAX = 9007199254740992.0;
+
     /** @var array Default values for options in the encoder */
     private static $defaultOptions = [
-        'float.integers' => false,
+        'float.integers'  => false,
         'float.precision' => 17,
     ];
 
@@ -34,39 +37,82 @@ class FloatEncoder implements Encoder
             return $value < 0 ? '-INF' : 'INF';
         }
 
-        return $this->getFloat($value, $options['float.precision'], $options['float.integers']);
+        return $this->encodeNumber($value, $options);
     }
 
     /**
-     * Converts the float value into string representation.
-     * @param float $float Value to convert
-     * @param integer|false $precision Number of decimals in the number or false for default
-     * @param boolean $useIntegers Whether to represent integer values as integers or not
-     * @return string The given float value as a string
+     * Encodes the number as a PHP number representation.
+     * @param float $float The number to encode
+     * @param array $options The float encoding options
+     * @return string The PHP code representation for the number
      */
-    private function getFloat($float, $precision, $useIntegers)
+    private function encodeNumber($float, array $options)
     {
-        if ($useIntegers && round($float) === $float) {
+        if ($this->isInteger($float, $options['float.integers'])) {
             return number_format($float, 0, '.', '');
-        } elseif ($precision === false) {
-            $output = (string) $float;
-        } else {
-            $original = ini_get('precision');
-            ini_set('precision', (int) $precision);
-            $output = (string) $float;
-            ini_set('precision', $original);
+        } elseif ($float === 0.0) {
+            return '0.0';
         }
 
-        return $this->enforceType($output);
+        $precision = $options['float.precision'];
+
+        if ($precision === false) {
+            $precision = ini_get('serialize_precision');
+        }
+
+        return $this->encodeFloat($float, $precision);
     }
 
     /**
-     * Ensures that the float representation will be parsed as float value.
-     * @param string $string Float string representation
-     * @return string Float value as string with enforced type
+     * Tells if the number can be encoded as an integer.
+     * @param float $float The number to test
+     * @param bool|string $allowIntegers Whether integers should be allowed
+     * @return bool True if the number can be encoded as an integer, false if not
      */
-    private function enforceType($string)
+    private function isInteger($float, $allowIntegers)
     {
-        return $string . (preg_match('/^[-+]?\d+$/', $string) ? '.0' : '');
+        if (!$allowIntegers || round($float) !== $float) {
+            return false;
+        } elseif (abs($float) < self::FLOAT_MAX) {
+            return true;
+        }
+
+        return $allowIntegers === 'all';
+    }
+
+    /**
+     * Encodes the number using a floating point representation.
+     * @param float $float The number to encode
+     * @param int $precision The maximum precision of encoded floats
+     * @return string The PHP code representation for the number
+     */
+    private function encodeFloat($float, $precision)
+    {
+        $precision = max(1, (int) $precision);
+        $log = (int) floor(log(abs($float), 10));
+
+        if (abs($float) < self::FLOAT_MAX && $log > -5 && abs($log) < $precision) {
+            return $this->formatFloat($float, $precision - $log - 1);
+        }
+
+        // Deal with overflow that results from rounding
+        $log += (int) (round(abs($float) / pow(10, $log), $precision - 1) / 10);
+        $string = $this->formatFloat($float / pow(10, $log), $precision - 1);
+
+        return sprintf('%sE%+d', $string, $log);
+    }
+
+    /**
+     * Formats the number as a decimal number.
+     * @param float $float The number to format
+     * @param int $digits The maximum number of decimal digits
+     * @return string The number formatted as a decimal number
+     */
+    private function formatFloat($float, $digits)
+    {
+        $digits = max((int) $digits, 1);
+        $string = rtrim(number_format($float, $digits, '.', ''), '0');
+
+        return substr($string, -1) === '.' ? $string . '0' : $string;
     }
 }