+++ src/InvalidOptionException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Riimu\Kit\PHPEncoder;
+
+/**
+ * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
+ * @copyright Copyright (c) 2015, Riikka Kalliomäki
+ * @license http://opensource.org/licenses/mit-license.php MIT License
+ */
+class InvalidOptionException extends \InvalidArgumentException
+{
+
+}