+++ src/PHPEncoder.php
@@ -33,13 +33,18 @@ class PHPEncoder
     /**
      * Creates a new PHPEncoder instance.
      *
-     * Optionally, you can provide a list of default options to the constructor
-     * as an associative array. If you do not want to use the default list of
-     * encoders supplied with the library, you may also provide a list of
-     * value encoders to the constructor.
+     * The constructor allows you to provide the list of default encoding options
+     * used by the encoder. Note that if you are using custom value encoders, you
+     * must provide them in the constructor if you are providing options for them
+     * or otherwise the options will be considered invalid.
+     *
+     * Using the second parameter you can also provide a list of value encoders used
+     * by the encoder. If null is provided, the list of default value encoders will
+     * be used instead.
      *
      * @param array $options List of encoder options
-     * @param null|Encoder[] $encoders List of encoders to use or null for default
+     * @param null|Encoder[] $encoders List of encoders to use or null for defaults
+     * @throws InvalidOptionException If any of the encoder options are invalid
      */
     public function __construct(array $options = [], array $encoders = null)
     {
@@ -73,7 +78,7 @@ class PHPEncoder
      * first.
      *
      * @param Encoder\Encoder $encoder Encoder for encoding values
-     * @param boolean $prepend True to prepend the encoder to the list, false to add it as last
+     * @param bool $prepend True to prepend the encoder to the list, false to add it as last
      */
     public function addEncoder(Encoder\Encoder $encoder, $prepend = false)
     {
@@ -81,15 +86,15 @@ class PHPEncoder
     }
 
     /**
-     * Sets an encoder option value.
-     * @param string $option Name of the option value
+     * Sets the value for an encoder option.
+     * @param string $option Name of the option
      * @param mixed $value Value for the option
-     * @throws \InvalidArgumentException If the option name is invalid
+     * @throws InvalidOptionException If the provided encoder option is invalid
      */
     public function setOption($option, $value)
     {
         if (!$this->isValidOption($option)) {
-            throw new \InvalidArgumentException('Invalid option name: ' . $option);
+            throw new InvalidOptionException("Invalid encoder option '$option'");
         }
 
         $this->options[$option] = $value;
@@ -102,12 +107,12 @@ class PHPEncoder
      */
     private function isValidOption($option)
     {
-        if (isset($this->options[$option])) {
+        if (array_key_exists($option, $this->options)) {
             return true;
         }
 
         foreach ($this->encoders as $encoder) {
-            if (isset($encoder->getDefaultOptions()[$option])) {
+            if (array_key_exists($option, $encoder->getDefaultOptions())) {
                 return true;
             }
         }
@@ -120,6 +125,9 @@ class PHPEncoder
      * @param mixed $variable Value to encode as PHP
      * @param array $options List of encoder options
      * @return string The PHP code that represents the given value
+     * @throws InvalidOptionException If any of the encoder options are invalid
+     * @throws \InvalidArgumentException If the provided value contains an unsupported value type
+     * @throws \RuntimeException If max depth is reached or a recursive value is detected
      */
     public function encode($variable, array $options = [])
     {
@@ -128,9 +136,9 @@ class PHPEncoder
 
     /**
      * Returns a list of all encoder options.
-     * @param array $overrides Options to override in the returned values
+     * @param array $overrides Options to override in the returned array
      * @return array List of encoder options
-     * @throws \InvalidArgumentException If any of the option overrides is invalid
+     * @throws InvalidOptionException If any of the encoder option overrides are invalid
      */
     public function getAllOptions(array $overrides = [])
     {
@@ -141,8 +149,8 @@ class PHPEncoder
         }
 
         foreach ($overrides as $name => $value) {
-            if (!isset($options[$name])) {
-                throw new \InvalidArgumentException('Invalid option name: ' . $name);
+            if (!array_key_exists($name, $options)) {
+                throw new InvalidOptionException("Invalid encoder option '$name'");
             }
 
             $options[$name] = $value;
@@ -155,11 +163,11 @@ class PHPEncoder
     /**
      * Generates the code for the given value recursively.
      * @param mixed $value Value to encode
-     * @param integer $depth Current indentation depth of the output
+     * @param int $depth Current indentation depth of the output
      * @param array $options List of encoder options
      * @param array $recursion Previously encoded values for recursion detection
      * @return string The PHP code that represents the given value
-     * @throws \RuntimeException On recursion or upon reaching maximum depth
+     * @throws \RuntimeException If max depth is reached or a recursive value is detected
      */
     private function generate($value, $depth, array $options, array $recursion = [])
     {
@@ -183,7 +191,8 @@ class PHPEncoder
      * @param mixed $value Value to try for circular reference
      * @param array $options List of encoder options
      * @param array $recursion Upper values in the encoding tree
-     * @return boolean True if values should be recorded, false if not
+     * @return bool True if values should be recorded, false if not
+     * @throws \RuntimeException If a recursive value is detected
      */
     private function detectRecursion(& $value, array $options, array $recursion)
     {
@@ -192,7 +201,7 @@ class PHPEncoder
                 if ($options['recursion.ignore']) {
                     $value = null;
                 } else {
-                    throw new \RuntimeException('Recursion detected');
+                    throw new \RuntimeException('A recursive value was detected');
                 }
             }
 
@@ -205,11 +214,11 @@ class PHPEncoder
     /**
      * Encodes the value using one of the encoders that supports the value type.
      * @param mixed $value Value to encode
-     * @param integer $depth Current indentation depth of the output
+     * @param int $depth Current indentation depth of the output
      * @param array $options List of encoder options
      * @param callable $encode Callback used to encode values
      * @return string The PHP code that represents the given value
-     * @throws \InvalidArgumentException If none of the encoders support the value type
+     * @throws \InvalidArgumentException If the provided value contains an unsupported value type
      */
     private function encodeValue($value, $depth, array $options, callable $encode)
     {
@@ -219,6 +228,6 @@ class PHPEncoder
             }
         }
 
-        throw new \InvalidArgumentException('Unable to encode value type: ' . gettype($value));
+        throw new \InvalidArgumentException(sprintf("Unsupported value type '%s'", gettype($value)));
     }
 }